+++ R/interface.R
@@ -87,6 +87,8 @@ RPhosFate <- function(...) {
 #' * `is_MCi`: An integer scalar specifying the current Monte Carlo iteration if
 #' applicable (defaults to `integer()`, which means Monte Carlo simulation mode
 #' is disabled).
+#' * `cv_MCl` A character vector specifying the names of the layers, which shall
+#' be written to hard disk for every Monte Carlo iteration.
 #'
 #' @section Model parameter arguments:
 #' * `ns_slp_min`: A numeric scalar specifying the minimum bounding slope in %
@@ -137,17 +139,11 @@ RPhosFate <- function(...) {
 #' This mode can make use of repeated random samples, i.e. raster data, of
 #' distributions of about all input data. The filenames of the Monte Carlo input
 #' data must contain the respective iteration, for example, _CFa12.tif_ for the
-#' twelfth iteration of the C-factors input data, and have to be put into a
+#' twelfth iteration of the C-factors input data, and have to reside in a
 #' separate directory. In case no Monte Carlo input file is found in the
 #' designated directory, the equivalent input data in the project directory is
-#' utilised. In order to save disk space, only the following model results are
-#' written to hard disk with the respective iteration in their filenames upon
-#' calling the appropriate methods (please note that some model results are not
-#' written/updated at all):
-#' * _ero_
-#' * _xxe_
-#' * _xxt_
-#' * \emph{xxt_cld}
+#' utilised. Please note that the loaded project may be in a inconsistent state
+#' after running one or more Monte Carlo iterations.
 #'
 #' @return An S4 [`RPhosFate-class`] river catchment object.
 #'